+++ app/models/tool.rb
@@ -9,7 +9,7 @@ class Tool < ActiveRecord::Base
 
   # Named Scopes
   scope :with_editor, lambda { |arg| where('tools.user_id IN (?) or tools.id in (select tool_users.tool_id from tool_users where tool_users.user_id = ? and tool_users.editor = ? and tool_users.approved = ?)', arg, arg, true, true ).references(:tool_users) }
-  scope :with_viewer, lambda { |arg| where('tools.user_id IN (?) or tools.public = ? or tools.id in (select tool_users.tool_id from tool_users where tool_users.user_id = ? and tool_users.approved = ?)', arg, true, arg, true ).references(:tool_users) }
+  scope :with_viewer_or_editor, lambda { |arg| where('tools.user_id IN (?) or tools.public = ? or tools.id in (select tool_users.tool_id from tool_users where tool_users.user_id = ? and tool_users.approved = ?)', arg, true, arg, true ).references(:tool_users) }
 
   # Model Validation
   validates_presence_of :name, :slug, :user_id