+++ CHANGELOG.md
@@ -67,6 +67,7 @@
 ### Bug Fix
 - Fixed a bug that caused too much information to be logged to the log file
 - Fixed an issue where friendly redirect on sign out was not redirecting to the last page the user was on
+- Dataset editors can now properly view private datasets on which they are an editor
 
 ## 0.15.3 (January 6, 2015)